重构并改进代码

+++ src/ObjectServer.Core/Data/IColumnMetadata.cs
@@ -13,7 +13,7 @@ namespace ObjectServer.Data
 
         string SqlType { get; }
 
-        long Length { get; }
+        int Length { get; }
 
         int Precision { get; }
     }
